requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
discord.py==1.7.3
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.2.0
future==0.18.2
idna==3.3
imageio-ffmpeg==0.4.5
multidict==5.2.0
numpy==1.22.0
py-cord==2.0.0a4748+gc602bc45
pycord==0.1.1
pycparser==2.21
pydub==0.25.1
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
